add appointment url and part of schedule

+++ database/migrations/2021_07_13_102105_create_log_events_table.php
@@ -17,8 +17,8 @@ class CreateLogEventsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->integer('code');
-            $table->integer('object');
+            $table->tinyInteger('code');
+            $table->tinyInteger('object');
             $table->timestamps();
         });
     }
